src/analysis/solver/sensitivity.ts
import { SolverState } from '../../../spec/Solver';
import { Matrix, createMatrix } from './matrix';

export interface SensitivityResult {
  shadowPrices: number[];
  reducedCosts: number[];
  allowableIncreases: number[];
  allowableDecreases: number[];
}

export function analyzeSensitivity(state: SolverState): SensitivityResult {
  // Calculate shadow prices (dual values)
  const shadowPrices = calculateShadowPrices(state);
  
  // Calculate reduced costs
  const reducedCosts = calculateReducedCosts(state);
  
  // Calculate allowable ranges
  const { increases, decreases } = calculateAllowableRanges(state);

  return {
    shadowPrices,
    reducedCosts,
    allowableIncreases: increases,
    allowableDecreases: decreases
  };
}

function calculateShadowPrices(state: SolverState): number[] {
  // Get the optimal basis
  const basis = getOptimalBasis(state);
  
  // Calculate dual values using the basis inverse
  const dualValues = calculateDualValues(basis);
  
  return dualValues;
}

function calculateReducedCosts(state: SolverState): number[] {
  // Get the optimal basis
  const basis = getOptimalBasis(state);
  
  // Calculate reduced costs for non-basic variables
  const reducedCosts = calculateNonBasicCosts(basis);
  
  return reducedCosts;
}

function calculateAllowableRanges(state: SolverState): {
  increases: number[];
  decreases: number[];
} {
  // Calculate range for each coefficient
  const increases: number[] = [];
  const decreases: number[] = [];
  
  // For each variable
  for (let i = 0; i < state.problem.variables.length; i++) {
    const { increase, decrease } = calculateCoefficientRange(state, i);
    increases.push(increase);
    decreases.push(decrease);
  }
  
  return { increases, decreases };
}

function getOptimalBasis(state: SolverState): Matrix {
  // Create coefficient matrix
  const matrix = createMatrix(
    state.problem.constraints.length,
    state.problem.variables.length
  );
  
  // Fill matrix with constraint coefficients
  // This is a placeholder implementation
  
  return matrix;
}

function calculateDualValues(basis: Matrix): number[] {
  // Calculate dual values using basis inverse
  // This is a placeholder implementation
  return [];
}

function calculateNonBasicCosts(basis: Matrix): number[] {
  // Calculate reduced costs for non-basic variables
  // This is a placeholder implementation
  return [];
}

function calculateCoefficientRange(state: SolverState, variableIndex: number): {
  increase: number;
  decrease: number;
} {
  // Calculate allowable range for coefficient
  // This is a placeholder implementation
  return {
    increase: 0,
    decrease: 0
  };
}
